refactor(svc-data): extract asset/portfolio access control services
