+++ svc-data/src/main/kotlin/com/beancounter/marketdata/assets/PrivateAssetConfigService.kt
@@ -3,9 +3,6 @@ package com.beancounter.marketdata.assets
 import com.beancounter.common.exception.BusinessException
 import com.beancounter.common.exception.ForbiddenException
 import com.beancounter.common.exception.NotFoundException
-import com.beancounter.common.model.ShareStatus
-import com.beancounter.common.model.SystemUser
-import com.beancounter.marketdata.portfolio.PortfolioShareRepository
 import com.beancounter.marketdata.registration.SystemUserService
 import com.beancounter.marketdata.trn.TrnRepository
 import jakarta.transaction.Transactional
@@ -17,8 +14,8 @@ import java.time.LocalDate
 /**
  * Service for managing private asset configurations.
  *
- * Handles CRUD operations for asset-level income/expense settings,
- * ensuring users can only access configs for assets they own.
+ * Handles CRUD operations for asset-level income/expense settings.
+ * Ownership and shared-read checks are delegated to [AssetAccessControl].
  */
 @Service
 @Transactional
@@ -27,7 +24,7 @@ class PrivateAssetConfigService(
     private val assetRepository: AssetRepository,
     private val systemUserService: SystemUserService,
     private val trnRepository: TrnRepository,
-    private val portfolioShareRepository: PortfolioShareRepository
+    private val accessControl: AssetAccessControl
 ) {
     private val log = LoggerFactory.getLogger(this::class.java)
 
@@ -83,10 +80,10 @@ class PrivateAssetConfigService(
      *
      * Read access is allowed for either the asset owner OR a viewer of any
      * portfolio that holds transactions for the asset (active share). Writes
-     * remain owner-only via [verifyAssetOwnership].
+     * remain owner-only via [AssetAccessControl.verifyOwnership].
      */
     fun getConfig(assetId: String): PrivateAssetConfigResponse? {
-        verifyAssetReadAccess(assetId)
+        accessControl.verifyReadAccess(assetId)
         val config = configRepository.findById(assetId).orElse(null) ?: return null
         return PrivateAssetConfigResponse(config)
     }
@@ -98,7 +95,7 @@ class PrivateAssetConfigService(
         assetId: String,
         request: PrivateAssetConfigRequest
     ): PrivateAssetConfigResponse {
-        verifyAssetOwnership(assetId)
+        accessControl.verifyOwnership(assetId)
         logSaveRequest(assetId, request)
         validateSubAccountCodes(request.subAccounts)
         validateTaxTreatment(request)
@@ -379,7 +376,7 @@ class PrivateAssetConfigService(
      * Delete config for an asset.
      */
     fun deleteConfig(assetId: String) {
-        verifyAssetOwnership(assetId)
+        accessControl.verifyOwnership(assetId)
         if (!configRepository.existsById(assetId)) {
             throw NotFoundException("Config not found for asset: $assetId")
         }
@@ -393,63 +390,5 @@ class PrivateAssetConfigService(
     fun getConfigsForAssets(assetIds: Collection<String>): PrivateAssetConfigsResponse {
         val configs = configRepository.findByAssetIdIn(assetIds)
         return PrivateAssetConfigsResponse(configs)
-    }
-
-    /**
-     * Verify the current user owns the specified asset.
-     */
-    private fun verifyAssetOwnership(assetId: String) {
-        val user =
-            systemUserService.getActiveUser()
-                ?: throw BusinessException(SystemUserService.USER_NOT_AUTHENTICATED)
-
-        val asset =
-            assetRepository.findById(assetId).orElseThrow {
-                NotFoundException("Asset not found: $assetId")
-            }
-
-        if (asset.systemUser?.id != user.id) {
-            throw BusinessException("Asset not owned by current user")
-        }
-    }
-
-    /**
-     * Read-only counterpart to [verifyAssetOwnership]. Allows the asset
-     * owner OR any user with an ACTIVE portfolio share covering a portfolio
-     * that holds transactions for this asset. Write paths must still call
-     * [verifyAssetOwnership].
-     */
-    private fun verifyAssetReadAccess(assetId: String) {
-        val user =
-            systemUserService.getActiveUser()
-                ?: throw BusinessException(SystemUserService.USER_NOT_AUTHENTICATED)
-
-        val asset =
-            assetRepository.findById(assetId).orElseThrow {
-                NotFoundException("Asset not found: $assetId")
-            }
-
-        if (asset.systemUser?.id == user.id) return
-        if (hasSharedPortfolioReadAccess(assetId, user)) return
-        throw BusinessException("Asset not accessible to current user")
-    }
-
-    private fun hasSharedPortfolioReadAccess(
-        assetId: String,
-        user: SystemUser
-    ): Boolean {
-        val sharedPortfolioIds =
-            portfolioShareRepository
-                .findBySharedWithAndStatus(
-                    user,
-                    ShareStatus.ACTIVE
-                ).mapNotNull { it.portfolio?.id }
-                .toSet()
-        if (sharedPortfolioIds.isEmpty()) return false
-        val assetIdsInSharedPortfolios =
-            trnRepository
-                .findDistinctAssetIdsByPortfolioIds(sharedPortfolioIds)
-                .toSet()
-        return assetId in assetIdsInSharedPortfolios
     }
 }
