+++ svc-data/src/test/kotlin/com/beancounter/marketdata/trn/TrnSettlementServiceTest.kt
@@ -49,6 +49,7 @@ class TrnSettlementServiceTest {
                 org.mockito.kotlin.mock(),
                 org.mockito.kotlin.mock(),
                 org.mockito.kotlin.mock(),
+                org.mockito.kotlin.mock(),
                 org.mockito.kotlin.mock()
             )
     }
